requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
langgraph
edge-tts
PyPDF2
typing-extensions
scipy
sounddevice
speechrecognition
openai
